user guide steps to configure the LVM partition size. (#191)
* user guide steps to configure the LVM partition size.

Signed-off-by: Srinivasamurthy, Ramakrishna <ramakrishna.srinivasamurthy@intel.com>

* Add subsections

Signed-off-by: Srinivasamurthy, Ramakrishna <ramakrishna.srinivasamurthy@intel.com>

* Update configure_edgenode_lvm_parition.rst

* Add configuration of LVM size during registration of edge node.

Signed-off-by: punam biswal <punam.biswal@intel.com>

* Update LVM partition configuration documentation

Expanded and clarified instructions for configuring rootfs and LVM partitions for persistent volumes on edge nodes. Added implementation details, assumptions, and requirements for single and multi-disk scenarios, and described integration with Kubernetes and OpenEBS CSI addon.

* Update LVM configuration list formatting in docs

Changed the implementation details and assumptions list in the LVM partition configuration guide from numbered format to a generic ordered list for improved readability and consistency.

* Add spacing for clarity in LVM configuration docs

Inserted a blank line for improved readability in the implementation details section of the edge node LVM partition configuration guide.

* Update LVM configuration instructions in user guide

Improved formatting and clarity in the section on configuring LVM size during edge node registration. Added a section heading and clarified instructions for using the `orch-cli` tool.

---------

Signed-off-by: Srinivasamurthy, Ramakrishna <ramakrishna.srinivasamurthy@intel.com>
Signed-off-by: punam biswal <punam.biswal@intel.com>
Co-authored-by: punam biswal <punam.biswal@intel.com>
Co-authored-by: Krishnamurthy Jambur <krishna.j.murthy@intel.com>
Co-authored-by: shankarsrinivas1 <shankar.srinivas.jayanthi@intel.com>