Update versioning.rst
Added a note atop the file to clarify versioning, and I tried to write it in a timeless way so that it won't have to be updated again. That said, please fact check the note and its links, and feel free to edit the note directly if need be. Thanks.